fully working balancer pool cache - just need to fix the probably broken tests
