+++ shared/src/balancer.rs
@@ -28,7 +28,6 @@
 pub mod event_handler;
 mod info_fetching;
 mod logexpmath;
+pub mod pool_cache;
 pub mod pool_fetching;
 mod pool_storage;
-pub mod pool_cache;
-pub mod registered_pool_fetching;
