+++ shared/src/balancer/event_handler.rs
@@ -30,7 +30,7 @@ use contracts::{
     balancer_v2_weighted_pool_factory::{self, Event as WeightedPoolFactoryEvent},
     BalancerV2WeightedPool2TokensFactory, BalancerV2WeightedPoolFactory,
 };
-use ethcontract::{common::DeploymentInformation, Event as EthContractEvent};
+use ethcontract::{common::DeploymentInformation, Event as EthContractEvent, H256};
 use model::TokenPair;
 use std::sync::Arc;
 use std::{collections::HashSet, ops::RangeInclusive};
@@ -88,8 +88,8 @@ impl BalancerPoolRegistry {
     pub async fn get_pools_containing_token_pairs(
         &self,
         token_pairs: HashSet<TokenPair>,
-    ) -> Vec<RegisteredWeightedPool> {
-        let mut pool_set_1 = self
+    ) -> HashSet<H256> {
+        let pool_set_1 = self
             .weighted_pool_updater
             .lock()
             .await
@@ -101,6 +101,22 @@ impl BalancerPoolRegistry {
             .await
             .store
             .pools_containing_token_pairs(token_pairs);
+        pool_set_1.union(&pool_set_2).copied().collect()
+    }
+
+    pub async fn get_pools(&self, pool_ids: &HashSet<H256>) -> Vec<RegisteredWeightedPool> {
+        let mut pool_set_1 = self
+            .weighted_pool_updater
+            .lock()
+            .await
+            .store
+            .pools_for(pool_ids);
+        let pool_set_2 = self
+            .two_token_pool_updater
+            .lock()
+            .await
+            .store
+            .pools_for(pool_ids);
         pool_set_1.extend(pool_set_2);
         pool_set_1
     }
